Comment out stuff to get core-frontend compiling.
